[CHG] (render-config) alternative setup_database
Dropping and Creating the database using psql and SQL strings instead
of client applications, because Render is complaining about connection
to the server with dropdb failing.

It may be because I'm already connecting to the database using this
port. I may need to specify a different port to connect via.

I might also need to change the DATABASE_URL environment variable, to
be the external connection string instead of the internal one, though
in theory the internal connection string should work.